Fix a possible overflow issu
Summary: Ref T5211

Test Plan: unit test

Reviewers: heliangliang, fenghonghua, liushaohui, cuijianwei, chengqiming, zhangguanghao1, zhangchen5

Reviewed By: zhangchen5

Maniphest Tasks: T5211

Differential Revision: https://phabricator.d.xiaomi.net/D23977

+++ hadoop-hdfs-project/hadoop-hdfs/src/contrib/raid/src/main/java/org/apache/hadoop/contrib/raid/BlockCodec.java
@@ -548,7 +548,7 @@ public class BlockCodec {
       int blockIdx = groupNo * dataBlocksNum + i;
       if (blockIdx < totalBlockNum && dataErasures.get(blockIdx) == null) {
         try {
-          int pos = (int) (blockIdx * blockSize + offset + roundNo * stripSize);
+          long pos = blockIdx * blockSize + offset + roundNo * stripSize;
           if (blockIdx == totalBlockNum - 1 && lastBlockLen != blockSize) {
             // Process the last block whose length is smaller than block size
             if (pos + data[i].length < fileStatus.getLen()) {
